backend/pyproject.toml
[tool.poetry]
name = "libremore-backend"
version = "0.1.0"
description = "City-wide library recommendation service backend"
authors = ["LibreMore Team <team@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.30"
psycopg = {extras = ["binary"], version = "^3.1.18"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
redis = "^5.0.4"
httpx = "^0.27.0"
feedparser = "^6.0.11"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = {extras = ["cli"], version = "^0.27.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
